streamlit-app/requirements.txt
streamlit==1.28.1
requests==2.31.0
plotly==5.17.0
pandas==2.1.3
python-dotenv==1.0.0
streamlit-chat==0.1.1